fix: update space creation steps (#1)
### Changes
The `createSpace` function in `client.js` has been enhanced to
streamline the space creation process. The new workflow now
automatically handles the following steps:
- Provisions the space.
- Saves the space.
- Creates a recovery account.
- Delegates access to the recovery account.

This update removes the need for manual steps, ensuring the recovery
account is set up as part of the space creation process.

> **Note:** This PR should only be merged after
[storacha/w3up#1546](https://github.com/storacha/w3up/pull/1546) has
been merged and released.